+++ src/main/java/kcs/funding/fundingboost/domain/security/repository/RefreshTokenRepository.java
@@ -1,6 +1,6 @@
 package kcs.funding.fundingboost.domain.security.repository;
 
-import kcs.funding.fundingboost.domain.entity.token.RefreshToken;
+import kcs.funding.fundingboost.domain.security.entity.token.RefreshToken;
 import org.springframework.data.repository.CrudRepository;
 
 public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
