+++ src/main/java/kcs/funding/fundingboost/domain/security/service/JwtAuthenticationService.java
@@ -12,9 +12,9 @@ import io.jsonwebtoken.security.Keys;
 import java.security.Key;
 import java.util.Date;
 import kcs.funding.fundingboost.domain.dto.response.login.JwtAccessTokenDto;
-import kcs.funding.fundingboost.domain.entity.token.RefreshToken;
 import kcs.funding.fundingboost.domain.exception.CommonException;
 import kcs.funding.fundingboost.domain.security.CustomUserDetails;
+import kcs.funding.fundingboost.domain.security.entity.token.RefreshToken;
 import kcs.funding.fundingboost.domain.security.repository.RefreshTokenRepository;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
